requirements.txt
pyyaml
flask
requests
gunicorn
